keys relocating
