+++ lib/keys/comment_types.dart
@@ -1,0 +1,6 @@
+part of 'package:mawa_package/mawa_package.dart';
+
+class CommentTypes {
+  static const String ticketComment = 'TICKETCOMMENT';
+  static const String leaveComment = 'LEAVEMANAGEMENT';
+}
