+++ lib/keys/field_option_types.dart
@@ -1,0 +1,56 @@
+part of 'package:mawa_package/mawa_package.dart';
+
+class FieldOptionTypes {
+  static const String fieldOptionDisputeReason = 'CLAIM-DISPUTE-REASON';
+  static const String fieldOptionCategory = 'CATEGORY';
+  static const String fieldOptionPriority = 'PRIORITY';
+  static const String fieldOptionIdType = 'ID-TYPE';
+  static const String fieldOptionGender = 'GENDER';
+  static const String fieldOptionTitle = 'TITLE';
+  static const String fieldOptionMaritalStatus = 'MARITAL-STATUS';
+  static const String fieldOptionSuburb = 'SUBURB';
+  static const String fieldOptionTown = 'TOWN';
+  static const String fieldOptionCity = 'CITY';
+  static const String fieldOptionContactType = 'CONTACT-TYPE';
+  static const String fieldAddress = 'ADDRESSES';
+  static const String fieldOptionAddressType = 'ADDRESS-TYPE';
+  static const String fieldPartnerRole = 'PARTNER-ROLE';
+  static const String fieldClaimDeclineReason = 'CLAIM-DECLINE-REASON';
+  static const String fieldPaymentMethod = 'PAYMENT-METHOD';
+  static const String fieldOptionPaymentRequestStatusReason =
+      'PAYMENT-REQUEST-STATUS-REASON';
+  static const String fieldPaymentRequestType = 'PAYMENT-REQUEST-TYPE';
+  static const String fieldBankAccountType = 'BANK-ACCOUNT-TYPE';
+  static const String fieldPaymentReason = 'PAYMENT-REASON';
+  static const String fieldBankName = 'BANK-NAME';
+  static const String fieldBranch = 'BRANCH';
+  static const String fieldDepartment = 'DEPARTMENT';
+  static const String fieldEmploymentType = 'EMPLOYMENT-TYPE';
+  static const String fieldJobTitle = 'JOB-TITLE';
+
+  static const String fieldClaimTypes = 'CLAIM-TYPE';
+  static const String fieldOptionBusinessIdentity = 'BUSINESS-IDENTITIES';
+
+  static const String productCategory = 'PRODUCT-CATEGORY';
+  static const String fieldOptionLayByStatus = 'LAYBY-STATUS';
+  static const String fieldOptionLayByStatusReason = 'LAYBY-STATUS-REASON';
+  static const String fieldOptionPartnerType = 'PARTNER-TYPE';
+  static const String uom = 'UOM';
+  static const String pricingType = 'PRICING-TYPE';
+  static const String membershipCreationType = 'MEMBERSHIP-CREATION-TYPE';
+  static const String suburb = 'SUBURB';
+  static const String userType = 'USER-TYPE';
+  static const String productAttribute = 'PRODUCT-ATTRIBUTE';
+  static const String membershipType = 'MEMBERSHIP-TYPE';
+  static const String salesArea = 'SALES-AREA';
+  static const String tenderType = 'TENDER-TYPE';
+  static const String documentType = 'DOCUMENT-TYPE';
+  static const String membershipStatuses = 'MEMBERSHIP-STATUSES';
+  static const String membershipStatusReasons = 'MEMBERSHIP-STATUS-REASONS';
+  static const String serviceRequestCategory = 'SERVICE-REQUEST-CATEGORY';
+  static const String serviceRequestDescription = 'SERVICE-REQUEST-DESCRIPTION';
+  static const String serviceRequestPriority = 'SERVICE-REQUEST-PRIORITY';
+  static const String serviceRequestStatus = 'SERVICE-REQUEST-STATUS-REASON';
+  static const String serviceRequestStatusReason = 'SERVICE-REQUEST-STATUS';
+
+}
