+++ lib/keys/json_payloads.dart
@@ -1,0 +1,250 @@
+part of 'package:mawa_package/mawa_package.dart';
+
+class JsonPayloads {
+  static const String suburb = 'suburb';
+  static const String city = 'city';
+  static const String town = 'town';
+  static const String province = 'province';
+
+  static const String claimId = 'claimId';
+  static const String comments = 'comments';
+  static const String id = 'id';
+  static const String filter = 'filter';
+  static const String filterValue = 'x';
+  static const String assignedTo = 'assignedTo';
+  static const String assignedToID = 'assignedToID';
+  static const String clintID = 'clintID';
+  static const String status = 'status';
+
+  static const String partnerID = 'partnerID';
+  static const String ticketID = 'ticketID';
+
+  static const String value = 'value';
+  static const String type = 'type';
+  static const String transaction = 'transaction';
+
+  static const String otp = 'otp';
+  static const String partnerEmail = 'partnerEmail';
+
+  static const String number = 'number';
+  static const String New = 'New';
+  static const String Open = 'Open';
+  static const String Closed = 'Closed';
+  static const String InProgress = 'Inprogress';
+  static const String Completed = 'Completed';
+  static const String Resolved = 'Resolved';
+
+  static const String loggedByID = 'loggedByID';
+  static const String approverID = 'approverID';
+  static const String startDate = 'startDate';
+  static const String endDate = 'endDate';
+  static const String description = 'description';
+  static const String subDescription = 'subDescription';
+  static const String leaveType = 'leaveType';
+
+  static const String documentType = 'documentType';
+  static const String parentType = 'parentType';
+  static const String parentReference = 'parentReference';
+
+  static const String fieldCode = 'code';
+  static const String fieldDescription = 'description';
+
+  static const String fieldOptionCode = 'code';
+  static const String fieldOptionDescription = 'description';
+
+  static const String field = 'field';
+
+  static const String workCenter = 'workcenter';
+  static const String position = 'position';
+
+  static const String workCenterDesc = 'description';
+  static const String workCenterPath = 'path';
+
+  static const String appName = 'appName';
+  static const String versionNumber = 'versionNumber';
+  static const String appUsable = 'appUsable';
+
+  static const String email = 'email';
+  static const String detail = 'detail';
+  static const String cellphone = 'cellphone';
+  static const String role = 'role';
+  static const String password = 'password';
+  static const String userType = 'userType';
+  static const String partner = 'partner';
+
+  static const String ticketDescription = 'description';
+  static const String ticketCategory = 'category';
+  static const String ticketSummary = 'subDescription';
+  static const String ticketPriority = 'priority';
+
+  static const String recievedBy = 'recievedBy';
+
+  static const String taskDescription = 'taskDescription';
+  static const String dueDate = 'dueDate';
+  static const String assignedBy = 'assignedBy';
+  static const String summary = 'summary';
+
+  static const String sms = 'sms';
+
+  static const String ticketTaskID = 'ticketTaskID';
+
+  static const String firstName = 'firstName';
+  static const String middleName = 'middleName';
+  static const String lastName = 'lastName';
+  static const String idNumber = 'idNumber';
+  static const String personIdNumber = 'idNumber';
+  static const String idType = 'idType';
+  static const String addressline1 = 'line1';
+  static const String addressline2 = 'line2';
+  static const String addressline3 = 'line3';
+  static const String addressline4 = 'line4';
+  static const String birthDate = 'birthDate';
+  static const String postalCode = 'postalCode';
+  static const String cellNumber = 'detail';
+  static const String gender = 'gender';
+  static const String title = 'title';
+  static const String maritalStatus = 'maritalStatus';
+
+  static const String validTo = 'validTo';
+  static const String attribute = 'attribute';
+  static const String validFrom = 'validFrom';
+  static const String priceType = 'priceType';
+  static const String priceTypeDescription = 'priceTypeDescription';
+  static const String notes = 'notes';
+  static const String declineReason = 'declineReason';
+
+  static const String productCategory = 'category';
+  static const String productDescription = 'description';
+  static const String quantity = 'quantity';
+
+  static const String deceasedId = 'deceasedId';
+  static const String claimantId = 'claimantId';
+  static const String bankAccount = 'bankAccount';
+  static const String bankName = 'bankName';
+  static const String branchName = 'branchName';
+  static const String branchCode = 'branchCode';
+  static const String accountType = 'accountType';
+  static const String accountNumber = 'accountNumber';
+
+  static const String categoryDescription = 'categoryDescription';
+
+  //KEYS FOR THE NEW BACKEND
+  static const String customer = 'customer';
+  static const String items = 'items';
+  static const String product = 'product';
+  static const String unit = 'items';
+  static const String unitPrice = 'unitPrice';
+  static const String deliveryDate = 'deliveryDate';
+  static const String supplier = 'supplier';
+  static const String claimant = 'claimant';
+  static const String deceased = 'deceased';
+  static const String member = 'member';
+  static const String membership = 'membership';
+  static const String deathDate = 'deathDate';
+  static const String burialDate = 'burialDate';
+  static const String code = 'code';
+  static const String category = 'category';
+  static const String baseUnitOfMeasure = 'baseUnitOfMeasure';
+  static const String sellingPrice = 'sellingPrice';
+
+  static const String customerId = 'customerId';
+  static const String expiryDate = 'expiryDate';
+  static const String productId = 'productId';
+  static const String ean = 'ean';
+  static const String uom = 'uom';
+  static const String lineTotal = 'lineTotal';
+
+  static const String username = 'username';
+  static const String amountDeposited = 'amountDeposited';
+
+  static const String name1 = 'name1';
+  static const String name2 = 'name2';
+  static const String name3 = 'name3';
+  static const String name4 = 'name4';
+
+  static const String employeeResponsibleId = 'employeeResponsibleId';
+  static const String reference = 'reference';
+  static const String amount = 'amount';
+  static const String bankDetails = 'bankDetails';
+
+  static const String file = 'file';
+  static const String uploadDate = 'uploadDate';
+  static const String uploadTime = 'uploadTime';
+  static const String uploadedBy = 'uploadedBy';
+  static const String downloadDate = 'downloadDate';
+  static const String downloadedBy = 'downloadedBy';
+  static const String reason = 'reason';
+
+  static const String previousProduct = 'previousProduct';
+  static const String statusReason = 'statusReason';
+  static const String salesRepresentativeId = 'salesRepresentativeId';
+  static const String premium = 'premium';
+
+  static const String line1 = 'line1';
+  static const String line2 = 'line2';
+  static const String line3 = 'line3';
+  static const String line4 = 'line4';
+
+  static const String employeeId = 'employeeId';
+  static const String branch = 'branch';
+  static const String groupID = 'groupID';
+  static const String department = 'department';
+  static const String identityDtos = 'identityDtos';
+  static const String addressDtos = 'addressDtos';
+  static const String contactDtos = 'contactDtos';
+
+  static const String partnerType = 'partnerType';
+  static const String surname = 'surname';
+  static const String organizationName = 'organizationName';
+
+  static const String memberId = 'memberId';
+  static const String membershipId = 'membershipId';
+  static const String dateOfBirth = 'dateOfBirth';
+  static const String membershipNumber = 'membershipNumber';
+  static const String receiptType = 'receiptType';
+  static const String invoiceNumber = 'invoiceNumber';
+  static const String membershipPeriod = 'membershipPeriod';
+  static const String tenderType = 'tenderType';
+  static const String salesRepresentative = 'salesRepresentative';
+  static const String salesArea = 'salesArea';
+  static const String dateJoined = 'dateJoined';
+  static const String price = 'price';
+  static const String pricingType = 'pricingType';
+  static const String transactionIdLink = 'transactionIdLink';
+  static const String partnerId = 'partnerId';
+  static const String accountHolder = 'accountHolder';
+  static const String layById = 'layById';
+  static const String fileType = 'fileType';
+  static const String autoGenerateCode = 'autoGenerateCode';
+  static const String creationType = 'creationType';
+  static const String previousInsurerId = 'previousInsurerId';
+  static const String pricing = 'pricing';
+  static const String lastReceiptDate = 'lastReceiptDate';
+  static const String membershipType = 'membershipType';
+  static const String recipientId = 'recipientId';
+  static const String paymentReason = 'paymentReason';
+  static const String paymentMethod = 'paymentMethod';
+  static const String location = 'location';
+  static const String terminalId = 'terminalId';
+  static const String objectType = 'objectType';
+  static const String objectId = 'objectId';
+  static const String attributeValue = 'value';
+  static const String bookDate = "bookDate";
+  static const String bookTime = "bookTime";
+  static const String receipts = 'receipts';
+  static const String dateAdded = 'dateAdded';
+  static const String dateEffective = 'dateEffective';
+  static const String language = 'language';
+
+  static const String processor = 'processor';
+  static const String recipient = 'recipient';
+  static const String subType = 'subType';
+  static const String extension = 'extension';
+  static const String plannedStartDate = 'plannedStartDate';
+  static const String plannedEndDate = 'plannedEndDate';
+  static const String parentId = 'parentId';
+  static const String actualStartDate = 'actualStartDate';
+  static const String actualEndDate = 'actualEndDate';
+
+// static const String  = '';
+}
