+++ lib/keys/json_responses.dart
@@ -1,0 +1,410 @@
+part of 'package:mawa_package/mawa_package.dart';
+
+class JsonResponses {
+  static const String employeeId = 'employeeId';
+  static const String amount = 'amount';
+  static const String assignedBy = 'assignedBy';
+  static const String assignedByID = 'assignedByID';
+  static const String assignedTo = 'assignedTo';
+  static const String assignedToID = 'assignedToID';
+
+  static const String category = 'category';
+  static const String checkoutId = 'checkoutId';
+  static const String clientId = 'clientId';
+
+  static const String dateLogged = 'dateLogged';
+  static const String description = 'description';
+  static const String dueDate = 'dueDate';
+  static const String duration = 'duration';
+  static const String cause = 'cause';
+  static const String message = 'message';
+
+  static const String field = 'field';
+
+  static const String firstName = 'firstName';
+  static const String genderTypeDescription = 'genderTypeDescription';
+
+  static const String id = 'id';
+
+  static const String lastName = 'lastName';
+
+  static const String middleName = 'middleName';
+
+  static const String period = 'period';
+  static const String priority = 'priority';
+  static const String processedBy = 'processedBy';
+  static const String processingDate = 'processingDate';
+  static const String processingTime = 'processingTime';
+
+  static const String reference = 'reference';
+  static const String refreshToken = 'refreshToken';
+
+  static const String status = 'status';
+  static const String subDescription = 'subDescription';
+
+  static const String token = 'token';
+
+  static const String username = 'username';
+  static const String userType = 'userType';
+
+  static const String parent = 'parent';
+  static const String fileName = 'fileName';
+  static const String extension = 'extension';
+  static const String createdBy = 'createdBy';
+  static const String createdAt = 'createdAt';
+  static const String attachedById = 'attachedById';
+  static const String attachedBy = 'attachedBy';
+
+  // static const String = '';
+
+  static const String fieldOptionCode = 'code';
+  static const String fieldOptionDescription = 'description';
+  static const String fieldOptionValidFrom = 'validFrom';
+  static const String fieldOptionValidTo = 'validTo';
+
+  static const String fieldCode = 'code';
+  static const String fieldDescription = 'description';
+  static const String fieldValidFrom = 'validFrom';
+  static const String fieldValidTo = 'validTo';
+
+  static const String usersFirstName = 'firstName';
+  static const String usersLastName = 'lastName';
+  static const String usersPartner = 'partner';
+  static const String userPartner = 'userPartner';
+  static const String cellphone = 'cellphone';
+  static const String email = 'email';
+  static const String usersStatus = 'status';
+  static const String usersGroupId = 'groupID';
+  static const String usersGroupName = 'groupName';
+
+  static const String usersRolesDescription = 'description';
+
+  static const String loginUserId = 'userID';
+  static const String loginPassword = 'password';
+
+  static const String cashupId = 'id';
+  static const String cashupAmount = 'amount';
+  static const String cashupStatus = 'status';
+  static const String cashupAmountDeposited = 'amountDeposited';
+  static const String cashupCashierID = 'cashierID';
+  static const String cashupCashier = 'cashier';
+  static const String cashupDateCreated = 'dateCreated';
+
+  static const String addressesType = 'type';
+  static const String addressesLine1 = 'line1';
+  static const String addressesLine2 = 'line2';
+  static const String addressesLine3 = 'line3';
+  static const String addressesLine4 = 'line4';
+
+  static const String line1 = 'line1';
+  static const String line2 = 'line2';
+  static const String line3 = 'line3';
+  static const String line4 = 'line4';
+  static const String postalCode = 'postalCode';
+
+  static const String address = 'address';
+  static const String addressesPostalCode = 'postalCode';
+  static const String addressesTypeDescription = 'typeDescription';
+  static const String addressline1Description = 'line1Description';
+  static const String addressline2Description = 'line2Description';
+  static const String addressline3Description = 'line3Description';
+  static const String adressline4Description = 'line4Description';
+  static const String addressline4Description = 'line4Description';
+
+  static const String maritalStatusDescription = 'maritalStatusDescription';
+
+  static const String addressPostalCode = 'postalcode';
+
+  static const String contactsType = 'type';
+  static const String contactTypeDescription = 'typeDescription';
+  static const String contactsDetail = 'detail';
+
+  static const String appName = 'appName';
+  static const String appUsable = 'appUsable';
+  static const String versionNumber = 'versionNumber';
+  static const String valid_from = 'valid_from';
+  static const String valid_to = 'valid_to';
+
+  static const String workCenterId = 'id';
+  static const String workCenterDescription = 'description';
+  static const String workcenterDescription = 'workcenterDescription';
+  static const String workCenterPath = 'path';
+
+  static const String personIdType = 'idType';
+  static const String personIdNumber = 'idnumber';
+  static const String personLastName = 'lastName';
+  static const String personMiddleName = 'middleName';
+  static const String personFirstName = 'firstName';
+  static const String personGender = 'gender';
+  static const String personBirthDate = 'birthDate';
+  static const String personLanguage = 'language';
+  static const String personMaritalStatus = 'maritalStatus';
+  static const String personTitle = 'title';
+  static const String personStatus = 'status';
+  static const String processor = 'processor';
+
+  static const String employeeID = 'id';
+  static const String employeeFirstName = 'name1';
+  static const String employeeLastName = 'name2';
+  static const String employeeIdNumber = 'idNumber';
+
+  static const String identityTypeDescription = 'typeDescription';
+
+  static const String titleDescription = 'titleDescription';
+
+  static const String ticketLogID = 'ticketLogID';
+  static const String ticketLogTicketID = 'ticketID';
+  static const String ticketLogPartnerId = 'partnerID';
+  static const String ticketLogPartner = 'partner';
+  static const String ticketLogStartTime = 'startTime';
+  static const String ticketLogEndTime = 'endTime';
+  static const String ticketLogNotes = 'notes';
+
+  static const String leaveRequestID = 'leaveRequestID';
+  static const String requestID = 'requestId';
+  static const String leaveStartDate = 'startDate';
+  static const String leaveEndDate = 'endDate';
+  static const String leaveType = 'leaveType';
+  static const String leaveTypeDescription = 'leaveTypeDescription';
+  static const String leaveStatus = 'status';
+  static const String leaveDateCreated = 'dateCreated';
+  static const String leaveUnitOfMeasure = 'unitOfMeasure';
+  static const String leaveQuantity = 'quantity';
+  static const String leaveQuantityRemaining = 'quantityRemaining';
+  static const String leaveStatusReason = 'statusReason';
+  static const String leaveLoggedByID = 'loggedByID';
+  static const String leaveLoggedBy = 'loggedBy';
+  static const String leaveApprovedBy = 'approvedBy';
+  static const String leaveApproverID = 'approverID';
+  static const String leaveDescription = 'description';
+
+  static const String leaveProfilesType = 'type';
+  static const String leaveProfilesLeaveType = 'leaveType';
+  static const String leaveProfilesStatus = 'status';
+  static const String leaveProfilesValidFrom = 'validFrom';
+  static const String leaveProfilesValidTo = 'validTo';
+
+  static const String partnerTypeDescription = 'partnerTypeDescription';
+
+  static const String leaveTypeItem = 'itemNo';
+  static const String leaveTypeProductCode = 'productCode';
+  static const String leaveTypeProductDescription = 'productDescription';
+  static const String leaveTypeQuantity = 'quantity';
+  static const String leaveTypeUnitOfMeasure = 'unitOfMeasure';
+  static const String leaveTypeQuantityRemaining = 'quantityRemaining';
+  static const String leaveTypeValidTo = 'validTo';
+  static const String leaveTypeValidFrom = 'validFrom';
+
+  static const String transactionValue = 'value';
+  static const String transactionDateCreated = 'date_created';
+  static const String transactionRelation = 'transaction';
+  static const String transactionTimeCreated = 'timeCreated';
+  static const String transactionLoggedById = 'loggedById';
+  static const String transactionLoggedBy = 'loggedBy';
+
+  static const String policyCustomerId = 'customerId';
+  static const String policyProductId = 'productId';
+  static const String policyProductProduct = 'product';
+  static const String policyDateJoined = 'dateJoined';
+  static const String policyDateEffective = 'dateEffective';
+  static const String policyCustomerDetails = 'customerDetails';
+  static const String policySalesRepId = 'salesRepresentativeId';
+  static const String policyStatus = 'status';
+  static const String policyStatusReason = 'statusReason';
+  static const String amountDue = 'amountDue';
+  static const String policySalesRepresentativeDetails =
+      'salesRepresentativeDetails';
+
+  static const String productDescription = 'productDescription';
+  static const String productCategory = 'category';
+  static const String productUnitPrice = 'unitPrice';
+
+  static const String path = 'path';
+
+  static const String workCenter = 'workcenter';
+  static const String role = 'role';
+  static const String position = 'position';
+
+  // static const String = '';
+  static const String note = 'note';
+
+  static const String ticketTaskID = 'ticketTaskID';
+  static const String personAssignTo = 'personAssignTo';
+
+  static const String leaveApprover = 'approver';
+
+  static const String type = 'type';
+  static const String summary = 'summary';
+  static const String taskDescription = 'taskDescription';
+  static const String ticketID = 'ticketID';
+  static const String startTime = 'startTime';
+  static const String personAssignedBy = 'personAssignedBy';
+  static const String notes = 'notes';
+  static const String taskLogs = 'taskLogs';
+  static const String endTime = 'endTime';
+
+  static const String product = 'product';
+
+  static const String tombstoneDependents = 'tombstoneDependents';
+
+  static const String claimant = 'claimant';
+  static const String deceased = 'deceased';
+  static const String submissionDate = 'submissionDate';
+  static const String idnumber = 'idnumber';
+  static const String idNumber = 'idNumber';
+  static const String name2 = 'name2';
+  static const String name3 = 'name3';
+  static const String name4 = 'name4';
+  static const String birthDate = 'birthDate';
+  static const String validFrom = 'validFrom';
+  static const String validTo = 'validTo';
+
+  static const String dateCreated = 'dateCreated';
+  static const String branch = 'branch';
+  static const String department = 'department';
+
+  static const String location = 'location';
+  static const String terminalType = 'terminalType';
+  static const String terminalId = 'terminalId';
+
+  static const String amountDeposited = 'amountDeposited';
+  static const String amountCollected = 'amountCollected';
+
+  static const String attribute = 'attribute';
+  static const String value = 'value';
+
+  static const String priceType = 'priceType';
+  static const String priceTypeDescription = 'priceTypeDescription';
+  static const String declineReason = 'declineReason';
+  static const String policy = 'policy';
+  static const String policyHolder = 'policyHolder';
+
+  static const String categoryDescription = 'categoryDescription';
+
+  //KEYS FOR THE NEW BACKEND
+
+  static const String number = 'number';
+  static const String code = 'code';
+
+  static const String procurementType = 'type';
+  static const String supplierId = 'supplierId';
+  static const String quantity = 'quantity';
+  static const String uom = 'uom';
+  static const String unitPrice = 'unitPrice';
+  static const String lineTotal = 'lineTotal';
+
+  static const String deliveryDate = 'deliveryDate';
+  static const String invoiceDate = 'invoiceDate';
+  static const String expiryDate = 'expiryDate';
+  static const String productId = 'productId';
+
+  static const String totalExcVat = 'totalExcVat';
+  static const String totalIncVat = 'totalIncVat';
+  static const String discountAmount = 'discountAmount';
+  static const String discountPercentage = 'discountPercentage';
+  static const String VATAmount = 'VATAmount';
+  static const String VATPercentage = 'VATPercentage';
+  static const String items = 'items';
+  static const String defaultFunction = 'defaultFunction';
+
+  static const String tenderType = 'tenderType';
+  static const String receiptType = 'receiptType';
+  static const String membershipHolder = 'membershipHolder';
+  static const String creationDate = 'creationDate';
+  static const String partner = 'partner';
+
+  static const String membershipId = 'membershipId';
+  static const String deathDate = 'deathDate';
+  static const String burialDate = 'burialDate';
+  static const String changedBy = 'changedBy';
+  static const String member = 'member';
+
+  static const String receipts = 'receipts';
+  static const String createdDate = 'createdDate';
+
+  // static const String changedBy = 'changedBy';
+
+  static const String lastUpdated = 'lastUpdated';
+  static const String statusReason = 'statusReason';
+  static const String memberId = 'memberId';
+  static const String deceasedId = 'deceasedId';
+
+  static const String title = 'title';
+  static const String receiptNumber = 'receiptNumber';
+  static const String name1 = 'name1';
+  static const String paymentReason = 'paymentReason';
+  static const String employeeResponsibleId = 'employeeResponsibleId';
+  static const String accountHolder = 'accountHolder';
+  static const String bankName = 'bankName';
+  static const String accountType = 'accountType';
+  static const String accountNumber = 'accountNumber';
+  static const String bankDetails = 'bankDetails';
+
+  static const typeDescription = 'typeDescription';
+  static const addressId = 'addressId';
+  static const String idType = 'idType';
+  static const String premium = 'premium';
+
+  static const String groupName = 'groupName';
+  static const String startDate = 'startDate';
+  static const String endDate = 'endDate';
+
+  static const String fieldClaimTypes = 'fieldClaimTypes';
+  static const String mainMember = 'mainMember';
+  static const String dependentDtoList = 'dependentDtoList';
+  static const String salesRep = 'salesRep';
+  static const String productDetails = 'productDetails';
+  static const String membershipNumber = 'membershipNumber';
+  static const String invoiceNumber = 'invoiceNumber';
+  static const String membershipPeriod = 'membershipPeriod';
+  static const String creationTime = 'creationTime';
+  static const String passwordStatus = 'passwordStatus';
+  static const String dateJoined = 'dateJoined';
+  static const String partnerId = 'partnerId';
+
+  //CLAIM
+
+  static const String branchCode = 'branchCode';
+  static const String claimantId = 'claimantId';
+  static const String baseUnitOfMeasure = 'baseUnitOfMeasure';
+  static const String sellingPrice = 'sellingPrice';
+  static const String customer = 'customer';
+  static const String salesRepresentative = 'salesRepresentative';
+  static const String createdOn = 'createdOn';
+  static const String price = 'price';
+  static const String surname = 'surname';
+  static const String tombstone = 'tombstone';
+  static const String fileType = 'fileType';
+  static const String no = 'no';
+  static const String fileId = 'fileId';
+  static const String file = 'file';
+  static const String membership = 'membership';
+  static const String paymentMethodDescription = 'paymentMethodDescription';
+  static const String statusDescription = 'statusDescription';
+
+  static const String employee = 'employee';
+
+  static const String paymentMethod = 'paymentMethod';
+  static const String recipientId = 'recipientId';
+  static const String pricing = 'pricing';
+  static const String gender = 'gender';
+  static const String pricingDescription = 'pricingDescription';
+  static const String documentType = 'documentType';
+  static const String maritalStatus = 'maritalStatus';
+  static const String identity = 'identity';
+  static const String baseUnitOfMeasureDescription =
+      'baseUnitOfMeasureDescription';
+  static const String uploadDate = 'uploadDate';
+  static const String claims = 'claims';
+  static const String amounts = 'amounts';
+
+  //Notification
+  static const String notificationDtos = 'notificationDtos';
+  static const String SubType = 'SubType';
+
+  //Bank
+  static const String partnerBankAccountDtoList = 'partnerBankAccountDtoList';
+
+  static const String recipient = 'recipient';
+  static const String employeeResponsible = 'employeeResponsible';
+}
