+++ lib/keys/membership_types.dart
@@ -1,0 +1,6 @@
+part of 'package:mawa_package/mawa_package.dart';
+
+class MembershipTypes {
+  static const String group = 'GROUP';
+  static const String individual = 'INDIVIDUAL';
+}
