+++ lib/keys/object_types.dart
@@ -1,0 +1,6 @@
+part of 'package:mawa_package/mawa_package.dart';
+
+class ObjectTypes {
+  static const String transaction = 'TRANSACTION';
+  static const String partner = 'PARTNER';
+}
