+++ lib/keys/partner_attributes.dart
@@ -1,0 +1,8 @@
+part of 'package:mawa_package/mawa_package.dart';
+
+class PartnerAttributes {
+  static const String identity = 'identity';
+  static const String contact = 'contact';
+  static const String address = 'address';
+  static const String bank = 'bank';
+}
