+++ lib/keys/partner_roles.dart
@@ -1,0 +1,8 @@
+part of 'package:mawa_package/mawa_package.dart';
+
+class PartnerRoles {
+  static const String competitor = 'COMPETITOR';
+  static const String dependent = 'DEPENDENT';
+  static const String customer = 'CUSTOMER';
+  static const String employee = 'EMPLOYEE';
+}
