+++ lib/keys/partner_types.dart
@@ -1,0 +1,13 @@
+part of 'package:mawa_package/mawa_package.dart';
+
+class PartnerTypes {
+  static const String dependent = 'DEPENDENT';
+  static const String customer = 'CUSTOMER';
+  static const String employee = 'EMPLOYEE';
+  static const String supplier = 'SUPPLIER';
+  static const String organization = 'ORGANIZATION';
+  static const String organisation = 'ORGANISATION';
+  static const String person = 'PERSON';
+  static const String group = 'GROUP';
+  static const String individual = 'individual';
+}
