+++ lib/keys/path_names.dart
@@ -1,0 +1,11 @@
+part of 'package:mawa_package/mawa_package.dart';
+
+class PathNames {
+  static const String identityPath = 'Identity';
+  static const String rolePath = 'Role';
+  static const String contactPath = 'Contact';
+  static const String addressPath = 'Address';
+  static const String employmentPath = 'Employment Details';
+  static const String employeePath = 'Employee Details';
+  static const String customerPath = 'Customer Details';
+}
