+++ lib/keys/product_types.dart
@@ -1,0 +1,6 @@
+part of 'package:mawa_package/mawa_package.dart';
+
+class ProductTypes {
+  static const String groupSociety = 'GROUP-SOCIETY';
+  static const String membership = 'MEMBERSHIP';
+}
