+++ lib/keys/query_parameters.dart
@@ -1,0 +1,150 @@
+part of 'package:mawa_package/mawa_package.dart';
+
+class QueryParameters {
+  static const String transactionId = 'transactionId';
+
+  static const String versionAppName = 'appName';
+  static const String versionAppUsable = 'appUsable';
+  static const String approverId = 'approverId';
+  static const String assignedTo = 'assignedTo';
+  static const String assignedToID = 'assignedToID';
+
+  static const String category = 'category';
+  static const String clintID = 'clintID';
+
+  static const String documentType = 'documentType';
+
+  static const String endDAte = 'endDate';
+  static const String extension = 'extension';
+
+  static const String field = 'field';
+  static const String filter = 'filter';
+  static const String filterValue = 'x';
+  static const String userType = 'userType';
+
+  static const String id = 'id';
+
+  static const String loggedByID = 'loggedByID';
+  static const String loggedById = 'loggedById';
+
+  static const String personIdNumber = 'idumber';
+  static const String partnerRole = 'partnerRole';
+  static const String idNumber = 'idNumber';
+  static const String idType = 'idType';
+
+  static const String organisationId = 'organisationId';
+  static const String otp = 'otp';
+
+  static const String parentType = 'parentType';
+  static const String parentReference = 'parentReference';
+  static const String partnerFunctionEmployee = 'EMPLOYEE';
+  static const String partnerFunctionOrganization = 'ORGANIZATION';
+  static const String partnerFunction = 'partnerFunction';
+  static const String partnerID = 'partnerID';
+  static const String partnerId = 'partnerId';
+  static const String partnerNo = 'partnerNo';
+  static const String processedBy = 'processedBy';
+  static const String processor = 'processor';
+
+  static const String reference = 'reference';
+
+  static const String status = 'status';
+
+  static const String ticketID = 'ticketID';
+  static const String transaction = 'transaction';
+
+  static const String versionApkVersionCode = 'versionNumber';
+  static const String value = 'value';
+
+  static const String meesageType = 'meesageType';
+  static const String adminEmail = 'adminEmail';
+  static const String user = 'user';
+  static const String password = 'password';
+
+  static const String assignTo = 'assignTo';
+  static const String ticketTaskID = 'ticketTaskID';
+  static const String summary = 'summary';
+  static const String taskDescription = 'taskDescription';
+  static const String dueDate = 'dueDate';
+  static const String taskLogID = 'taskLogID';
+
+  static const String userID = 'userID';
+  static const String note = 'note';
+  static const String type = 'type';
+
+  static const String workcenter = 'workcenter';
+  static const String position = 'position';
+
+  static const String New = 'new';
+  static const String inprogress = 'inprogress';
+  static const String open = 'open';
+  static const String cancelled = 'cancelled';
+  static const String resolved = 'resolved';
+  static const String completed = 'completed';
+
+  static const String organizationId = 'organizationId';
+
+  static const String policy = 'policy';
+  static const String approver = 'approver';
+
+  static const String salesArea = 'salesArea';
+  static const String product = 'product';
+  static const String customer = 'customer';
+  static const String salesRepresentative = 'salesRepresentative';
+
+  static const String amount = 'amount';
+  static const String location = 'location';
+  static const String terminalType = 'terminalType';
+  static const String terminalId = 'terminalId';
+  static const String tenderType = 'tenderType';
+  static const String transactionLink = 'transactionLink';
+  static const String role = 'role';
+
+  static const String cashupId = 'cashupId';
+  static const String checkoutId = 'checkoutId';
+  static const String statusReason = 'statusReason';
+  static const String description = 'description';
+
+  static const String email = 'email';
+  static const String cellphone = 'cellphone';
+  static const String passwordStatus = 'passwordStatus';
+  static const String userRole = 'userRole';
+
+  static const String fieldOption = 'fieldOption';
+
+  static const String addressId = 'addressId';
+  static const String startDate = 'startDate';
+  static const String createdBy = 'createdBy';
+  static const String createdOn = 'createdOn';
+  static const String customerId = 'customerId';
+  static const String endDate = 'endDate';
+  static const String number = 'number';
+
+  static const String membershipPeriod = 'membershipPeriod';
+  static const String membershipNumber = 'membershipNumber';
+  static const String invoiceNumber = 'invoiceNumber';
+  static const String receiptType = 'receiptType';
+  static const String transactionLinkId = 'transactionLinkId';
+  static const String fileType = 'fileType';
+  static const String claimant = 'claimant';
+  static const String no = 'no';
+  static const String deceased = 'deceased';
+  static const String member = 'member';
+  static const String membership = 'membership';
+  static const String deathDate = 'deathDate';
+  static const String burialDate = 'burialDate';
+
+  static const String contactType = 'contactType';
+  static const String attribute = 'attribute';
+  static const String notCashed = 'notCashed';
+  static const String pricing = 'pricing';
+  static const String employeeResponsibleId = 'employeeResponsibleId';
+  static const String receipts = 'receipts';
+  static const String employeeId = "employeeId";
+  static const String bookDate = "bookDate";
+  static const String attributeName = "attributeName";
+  static const String attributeValue = "attributeValue";
+  static const String objectId = "objectId";
+
+// static const String  = '';
+}
