+++ lib/keys/resources.dart
@@ -1,0 +1,171 @@
+part of 'package:mawa_package/mawa_package.dart';
+
+class Resources {
+  static const String address = 'address';
+  static const String bankAccount = 'bank-account';
+  static const String assign = 'assign';
+  static const String terminate = 'terminate';
+  static const String rehire = 'rehire';
+  static const String suspend = 'suspend';
+  static const String validateSAID = 'validateSAID';
+  static const String attachments = 'attachments';
+  static const String otp = 'otp';
+  static const String employee = 'employee';
+  static const String employees = 'employees';
+  static const String authenticate = 'authenticate';
+  static const String tickets = 'tickets';
+  static const String ticketsLog = 'ticketsLog';
+  static const String ticketInProgress = 'inprogress';
+  static const String ticketAwaiting = 'awaitingCustomer';
+  static const String ticketResolve = 'resolved';
+  static const String ticketReject = 'reject';
+  static const String reject = 'reject';
+  static const String dispute = 'dispute';
+  static const String ticketCancel = 'cancel';
+  static const String archive = 'archive';
+  static const String unarchive = 'unarchive';
+  static const String user = 'user';
+  static const String resetPassword = 'passwordreset';
+  static const String reset = 'reset';
+  static const String forgotPassword = 'forgotPassword';
+  static const String versions = 'versions';
+  static const String leaves = 'leaves';
+  static const String fields = 'fields';
+  static const String fieldOptions = 'fieldoptions';
+  static const String leavesApprovers = 'leavesApprovers';
+  static const String approversHistory = 'approversHistory';
+  static const String leavesToApprove = 'leavesToApprove';
+  static const String leaveProfiles = 'leaveProfiles';
+  static const String transactionNotes = 'transactionNotes';
+  static const String persons = 'persons';
+  static const String workcenter = 'workcenter';
+  static const String roleWorkcenters = 'roleWorkcenters';
+  static const String workCenterDesc = 'workcenterDesc';
+  static const String workCenterId = 'id';
+  static const String receipts = 'receipts';
+  static const String receipt = 'receipt';
+  static const String cashup = 'cashup';
+  static const String policies = 'policies';
+
+  static const String voucherValue = 'vouchers';
+  static const String identity = 'identity';
+  static const String contactValue = 'contacts';
+
+  static const String approve = 'approve';
+  static const String approveCancellation = 'ApproveCancellation';
+  static const String approveEditedLeave = 'ApproveEditedLeave';
+  static const String rejectApproval = 'RejectApproval';
+  static const String rejectCancel = 'RejectCancel';
+  static const String rejectEditedLeave = 'rejectEditedLeave';
+  static const String cancel = 'cancel';
+  static const String edit = 'Edit';
+  static const String roles = 'roles';
+  static const String role = 'role';
+  static const String products = 'products';
+  static const String editPosition = 'editPosition';
+  static const String addWorkcenter = 'addWorkcenter';
+  static const String removeWokcenter = 'removeWorkCenter';
+
+  static const String ticketStatusNew = 'new';
+  static const String ticketStatusOpen = 'open';
+  static const String closed = 'closed';
+  static const String inprogress = 'inprogress';
+  static const String awaitingCustomer = 'awaitingCustomer';
+
+  static const String addresses = 'addresses';
+  static const String sendNotifications = 'sendNotifications';
+
+  static const String getTasksAssigned = 'getTasksAssigned';
+  static const String editTask = 'editTask';
+  static const String startTask = 'startTask';
+  static const String cancelTask = 'cancelTask';
+  static const String endTask = 'endTask';
+  static const String completeTask = 'completeTask';
+  static const String getSpecificTask = 'getSpecificTask';
+  static const String getTasks = 'getTasks';
+
+  static const String refreshAuthenticate = 'refreshAuthenticate';
+  static const String comment = 'comment';
+
+  static const String createTask = 'createTask';
+
+  static const String organization = 'Organization';
+  static const String OrganizationUsers = 'OrganizationUsers';
+
+  static const String dependents = 'dependents';
+  static const String claims = 'claims';
+  static const String product = 'product';
+  static const String identities = 'identities';
+  static const String contacts = 'contacts';
+  static const String contact = 'contact';
+  static const String employeeId = 'employeeId';
+  static const String startDate = 'startDate';
+  static const String branch = 'branch';
+  static const String employment = 'employment';
+
+  static const String claimApprove = 'approve';
+  static const String claimCancel = 'cancel';
+  static const String claimDecline = 'decline';
+
+  static const String deposits = 'deposits';
+  static const String deposit = 'deposit';
+
+  static const String attributes = 'attributes';
+  static const String pricing = 'pricing';
+  static const String refer = 'refer';
+  static const String vouchers = 'vouchers';
+
+  static const String submit = 'submit';
+  static const String submitted = 'submitted';
+
+  //KEYS FOR THE NEW BACKEND
+
+  static const String purchaseOrder = 'purchase-order';
+  static const String salesOrder = 'sales-order';
+  static const String invoice = 'invoice';
+  static const String field = 'field';
+  static const String option = 'option';
+  static const String claim = 'claim';
+  static const String membership = 'membership';
+  static const String dependent = 'dependent';
+  static const String partner = 'partner';
+
+  static const String quotation = 'quotation';
+  static const String items = 'items';
+
+  static const String calculate_pricing = 'calculate-pricing';
+  static const String application = 'application';
+  static const String customer = 'customer';
+  static const String prospects = 'prospects';
+  static const String prospect = 'prospect';
+  static const String supplier = 'supplier';
+  static const String paymentRequest = 'payment-request';
+
+  static const String attachmentfile = 'attachmentfile';
+  static const String attachment = 'attachment';
+
+  static const String lock = 'lock';
+  static const String unlock = 'unlock';
+  static const String byId = 'byId';
+  static const String addressId = 'addressId';
+  static const String validate = 'validate';
+
+  static const String layBy = 'layBy';
+  static const String voucher = 'voucher';
+  static const String transactionLink = 'transactionLink';
+  static const String transaction = 'transaction';
+  static const String serviceRequest = 'service-request';
+  static const String tombstoneRecipient = 'tombstone-recipient';
+  static const String attribute = 'attribute';
+  static const String premium = 'premium';
+  static const String booking = 'booking';
+  static const String groupSociety = 'group-society';
+  static const String premiums = 'premiums';
+
+  static const String notifications = 'notifications';
+  static const String read = 'read';
+  static const String byTrans = 'byTrans';
+
+  static const String complete = 'complete';
+  static const String task = 'task';
+}
