+++ lib/keys/shared_preferences.dart
@@ -1,0 +1,23 @@
+part of 'package:mawa_package/mawa_package.dart';
+
+class SharedPrefs {
+  static const server = 'server';
+  static const token = 'token';
+  static const username = 'username';
+  static const lastPage = 'lastPage';
+  static const startTime = 'startTime';
+  static const endTime = 'endTime';
+  static const isTracking = 'isTracking';
+  static const trackingTicketID = 'trackingTicketID';
+  static const refreshToken = 'refreshToken';
+  static const initialRoute = 'initialRoute';
+  static const loginScreenID = 'loginScreenID';
+  static const loginRole = 'loginRole';
+  static const isLoggedIn = 'isLoggedIn';
+  static const tenantID = 'tenantID';
+  static const userID = 'userID';
+  static const role = 'role';
+  static const workcenters = 'workcenters';
+  static const loggedInUser = 'loggedInUser';
+  static const roleRoot = 'roleRoot';
+}
