+++ lib/keys/status_reasons.dart
@@ -1,0 +1,8 @@
+part of 'package:mawa_package/mawa_package.dart';
+
+class StatusReasons {
+  static const String awaitingApproval = 'Awaiting Approval';
+  static const String awaitingCancelApproval = 'Awaiting Cancel Approval';
+  static const String awaitingEditApproval = 'Awaiting Edit Approval';
+  static const String Rejected = 'Rejected';
+}
