+++ lib/keys/statuses.dart
@@ -1,0 +1,8 @@
+part of 'package:mawa_package/mawa_package.dart';
+
+class Statuses {
+  static const String pending = 'Pending';
+  static const String approved = 'Approved';
+  static const String cancelled = 'Cancelled';
+  static const String rejected = 'Rejected';
+}
