--- lib/services/keys.dart
@@ -1,1122 +1,0 @@
-part of 'package:mawa_package/mawa_package.dart';
-
-class JsonResponses {
-  static const String employeeId = 'employeeId';
-  static const String amount = 'amount';
-  static const String assignedBy = 'assignedBy';
-  static const String assignedByID = 'assignedByID';
-  static const String assignedTo = 'assignedTo';
-  static const String assignedToID = 'assignedToID';
-
-  static const String category = 'category';
-  static const String checkoutId = 'checkoutId';
-  static const String clientId = 'clientId';
-
-  static const String dateLogged = 'dateLogged';
-  static const String description = 'description';
-  static const String dueDate = 'dueDate';
-  static const String duration = 'duration';
-  static const String cause = 'cause';
-  static const String message = 'message';
-
-  static const String field = 'field';
-
-  static const String firstName = 'firstName';
-  static const String genderTypeDescription = 'genderTypeDescription';
-
-  static const String id = 'id';
-
-  static const String lastName = 'lastName';
-
-  static const String middleName = 'middleName';
-
-  static const String period = 'period';
-  static const String priority = 'priority';
-  static const String processedBy = 'processedBy';
-  static const String processingDate = 'processingDate';
-  static const String processingTime = 'processingTime';
-
-  static const String reference = 'reference';
-  static const String refreshToken = 'refreshToken';
-
-  static const String status = 'status';
-  static const String subDescription = 'subDescription';
-
-  static const String token = 'token';
-
-  static const String username = 'username';
-  static const String userType = 'userType';
-
-  static const String parent = 'parent';
-  static const String fileName = 'fileName';
-  static const String extension = 'extension';
-  static const String createdBy = 'createdBy';
-  static const String createdAt = 'createdAt';
-  static const String attachedById = 'attachedById';
-  static const String attachedBy = 'attachedBy';
-
-  // static const String = '';
-
-  static const String fieldOptionCode = 'code';
-  static const String fieldOptionDescription = 'description';
-  static const String fieldOptionValidFrom = 'validFrom';
-  static const String fieldOptionValidTo = 'validTo';
-
-  static const String fieldCode = 'code';
-  static const String fieldDescription = 'description';
-  static const String fieldValidFrom = 'validFrom';
-  static const String fieldValidTo = 'validTo';
-
-  static const String usersFirstName = 'firstName';
-  static const String usersLastName = 'lastName';
-  static const String usersPartner = 'partner';
-  static const String userPartner = 'userPartner';
-  static const String cellphone = 'cellphone';
-  static const String email = 'email';
-  static const String usersStatus = 'status';
-  static const String usersGroupId = 'groupID';
-  static const String usersGroupName = 'groupName';
-
-  static const String usersRolesDescription = 'description';
-
-  static const String loginUserId = 'userID';
-  static const String loginPassword = 'password';
-
-  static const String cashupId = 'id';
-  static const String cashupAmount = 'amount';
-  static const String cashupStatus = 'status';
-  static const String cashupAmountDeposited = 'amountDeposited';
-  static const String cashupCashierID = 'cashierID';
-  static const String cashupCashier = 'cashier';
-  static const String cashupDateCreated = 'dateCreated';
-
-  static const String addressesType = 'type';
-  static const String addressesLine1 = 'line1';
-  static const String addressesLine2 = 'line2';
-  static const String addressesLine3 = 'line3';
-  static const String addressesLine4 = 'line4';
-
-  static const String line1 = 'line1';
-  static const String line2 = 'line2';
-  static const String line3 = 'line3';
-  static const String line4 = 'line4';
-  static const String postalCode = 'postalCode';
-
-  static const String address = 'address';
-  static const String addressesPostalCode = 'postalCode';
-  static const String addressesTypeDescription = 'typeDescription';
-  static const String addressline1Description = 'line1Description';
-  static const String addressline2Description = 'line2Description';
-  static const String addressline3Description = 'line3Description';
-  static const String adressline4Description = 'line4Description';
-  static const String addressline4Description = 'line4Description';
-
-  static const String maritalStatusDescription = 'maritalStatusDescription';
-
-  static const String addressPostalCode = 'postalcode';
-
-  static const String contactsType = 'type';
-  static const String contactTypeDescription = 'typeDescription';
-  static const String contactsDetail = 'detail';
-
-  static const String appName = 'appName';
-  static const String appUsable = 'appUsable';
-  static const String versionNumber = 'versionNumber';
-  static const String valid_from = 'valid_from';
-  static const String valid_to = 'valid_to';
-
-  static const String workCenterId = 'id';
-  static const String workCenterDescription = 'description';
-  static const String workcenterDescription = 'workcenterDescription';
-  static const String workCenterPath = 'path';
-
-  static const String personIdType = 'idType';
-  static const String personIdNumber = 'idnumber';
-  static const String personLastName = 'lastName';
-  static const String personMiddleName = 'middleName';
-  static const String personFirstName = 'firstName';
-  static const String personGender = 'gender';
-  static const String personBirthDate = 'birthDate';
-  static const String personLanguage = 'language';
-  static const String personMaritalStatus = 'maritalStatus';
-  static const String personTitle = 'title';
-  static const String personStatus = 'status';
-  static const String processor = 'processor';
-
-  static const String employeeID = 'id';
-  static const String employeeFirstName = 'name1';
-  static const String employeeLastName = 'name2';
-  static const String employeeIdNumber = 'idNumber';
-
-  static const String identityTypeDescription = 'typeDescription';
-
-  static const String titleDescription = 'titleDescription';
-
-  static const String ticketLogID = 'ticketLogID';
-  static const String ticketLogTicketID = 'ticketID';
-  static const String ticketLogPartnerId = 'partnerID';
-  static const String ticketLogPartner = 'partner';
-  static const String ticketLogStartTime = 'startTime';
-  static const String ticketLogEndTime = 'endTime';
-  static const String ticketLogNotes = 'notes';
-
-  static const String leaveRequestID = 'leaveRequestID';
-  static const String requestID = 'requestId';
-  static const String leaveStartDate = 'startDate';
-  static const String leaveEndDate = 'endDate';
-  static const String leaveType = 'leaveType';
-  static const String leaveTypeDescription = 'leaveTypeDescription';
-  static const String leaveStatus = 'status';
-  static const String leaveDateCreated = 'dateCreated';
-  static const String leaveUnitOfMeasure = 'unitOfMeasure';
-  static const String leaveQuantity = 'quantity';
-  static const String leaveQuantityRemaining = 'quantityRemaining';
-  static const String leaveStatusReason = 'statusReason';
-  static const String leaveLoggedByID = 'loggedByID';
-  static const String leaveLoggedBy = 'loggedBy';
-  static const String leaveApprovedBy = 'approvedBy';
-  static const String leaveApproverID = 'approverID';
-  static const String leaveDescription = 'description';
-
-  static const String leaveProfilesType = 'type';
-  static const String leaveProfilesLeaveType = 'leaveType';
-  static const String leaveProfilesStatus = 'status';
-  static const String leaveProfilesValidFrom = 'validFrom';
-  static const String leaveProfilesValidTo = 'validTo';
-
-  static const String partnerTypeDescription = 'partnerTypeDescription';
-
-
-  static const String leaveTypeItem = 'itemNo';
-  static const String leaveTypeProductCode = 'productCode';
-  static const String leaveTypeProductDescription = 'productDescription';
-  static const String leaveTypeQuantity = 'quantity';
-  static const String leaveTypeUnitOfMeasure = 'unitOfMeasure';
-  static const String leaveTypeQuantityRemaining = 'quantityRemaining';
-  static const String leaveTypeValidTo = 'validTo';
-  static const String leaveTypeValidFrom = 'validFrom';
-
-  static const String transactionValue = 'value';
-  static const String transactionDateCreated = 'date_created';
-  static const String transactionRelation = 'transaction';
-  static const String transactionTimeCreated = 'timeCreated';
-  static const String transactionLoggedById = 'loggedById';
-  static const String transactionLoggedBy = 'loggedBy';
-
-  static const String policyCustomerId = 'customerId';
-  static const String policyProductId = 'productId';
-  static const String policyProductProduct = 'product';
-  static const String policyDateJoined = 'dateJoined';
-  static const String policyDateEffective = 'dateEffective';
-  static const String policyCustomerDetails = 'customerDetails';
-  static const String policySalesRepId = 'salesRepresentativeId';
-  static const String policyStatus = 'status';
-  static const String policyStatusReason = 'statusReason';
-  static const String amountDue = 'amountDue';
-  static const String policySalesRepresentativeDetails =
-      'salesRepresentativeDetails';
-
-  static const String productDescription = 'productDescription';
-  static const String productCategory = 'category';
-  static const String productUnitPrice = 'unitPrice';
-
-  static const String path = 'path';
-
-  static const String workCenter = 'workcenter';
-  static const String role = 'role';
-  static const String position = 'position';
-
-  // static const String = '';
-  static const String note = 'note';
-
-  static const String ticketTaskID = 'ticketTaskID';
-  static const String personAssignTo = 'personAssignTo';
-
-  static const String leaveApprover = 'approver';
-
-  static const String type = 'type';
-  static const String summary = 'summary';
-  static const String taskDescription = 'taskDescription';
-  static const String ticketID = 'ticketID';
-  static const String startTime = 'startTime';
-  static const String personAssignedBy = 'personAssignedBy';
-  static const String notes = 'notes';
-  static const String taskLogs = 'taskLogs';
-  static const String endTime = 'endTime';
-
-  static const String product = 'product';
-
-  static const String tombstoneDependents = 'tombstoneDependents';
-
-  static const String claimant = 'claimant';
-  static const String deceased = 'deceased';
-  static const String submissionDate = 'submissionDate';
-  static const String idnumber = 'idnumber';
-  static const String idNumber = 'idNumber';
-  static const String name2 = 'name2';
-  static const String name3 = 'name3';
-  static const String name4 = 'name4';
-  static const String birthDate = 'birthDate';
-  static const String validFrom = 'validFrom';
-  static const String validTo = 'validTo';
-
-  static const String dateCreated = 'dateCreated';
-  static const String branch = 'branch';
-  static const String department = 'department';
-
-  static const String location = 'location';
-  static const String terminalType = 'terminalType';
-  static const String terminalId = 'terminalId';
-
-  static const String amountDeposited = 'amountDeposited';
-  static const String amountCollected = 'amountCollected';
-
-  static const String attribute = 'attribute';
-  static const String value = 'value';
-
-  static const String priceType = 'priceType';
-  static const String priceTypeDescription = 'priceTypeDescription';
-  static const String declineReason = 'declineReason';
-  static const String policy = 'policy';
-  static const String policyHolder = 'policyHolder';
-
-  static const String categoryDescription = 'categoryDescription';
-
-  //KEYS FOR THE NEW BACKEND
-
-  static const String number = 'number';
-  static const String code = 'code';
-
-  static const String procurementType = 'type';
-  static const String supplierId = 'supplierId';
-  static const String quantity = 'quantity';
-  static const String uom = 'uom';
-  static const String unitPrice = 'unitPrice';
-  static const String lineTotal = 'lineTotal';
-
-  static const String deliveryDate = 'deliveryDate';
-  static const String invoiceDate = 'invoiceDate';
-  static const String expiryDate = 'expiryDate';
-  static const String productId = 'productId';
-
-  static const String totalExcVat = 'totalExcVat';
-  static const String totalIncVat = 'totalIncVat';
-  static const String discountAmount = 'discountAmount';
-  static const String discountPercentage = 'discountPercentage';
-  static const String VATAmount = 'VATAmount';
-  static const String VATPercentage = 'VATPercentage';
-  static const String items = 'items';
-  static const String defaultFunction = 'defaultFunction';
-
-  static const String tenderType = 'tenderType';
-  static const String receiptType = 'receiptType';
-  static const String membershipHolder = 'membershipHolder';
-  static const String creationDate = 'creationDate';
-  static const String partner = 'partner';
-
-  static const String membershipId = 'membershipId';
-  static const String deathDate = 'deathDate';
-  static const String burialDate = 'burialDate';
-  static const String changedBy = 'changedBy';
-  static const String member = 'member';
-
-  static const String receipts = 'receipts';
-  static const String createdDate = 'createdDate';
-
-  // static const String changedBy = 'changedBy';
-
-  static const String lastUpdated = 'lastUpdated';
-  static const String statusReason = 'statusReason';
-  static const String memberId = 'memberId';
-  static const String deceasedId = 'deceasedId';
-
-  static const String title = 'title';
-  static const String receiptNumber = 'receiptNumber';
-  static const String name1 = 'name1';
-  static const String paymentReason = 'paymentReason';
-  static const String employeeResponsibleId = 'employeeResponsibleId';
-  static const String accountHolder = 'accountHolder';
-  static const String bankName = 'bankName';
-  static const String accountType = 'accountType';
-  static const String accountNumber = 'accountNumber';
-  static const String bankDetails = 'bankDetails';
-
-  static const typeDescription = 'typeDescription';
-  static const addressId = 'addressId';
-  static const String idType = 'idType';
-  static const String premium = 'premium';
-
-  static const String groupName = 'groupName';
-  static const String startDate = 'startDate';
-  static const String endDate = 'endDate';
-
-  static const String fieldClaimTypes = 'fieldClaimTypes';
-  static const String mainMember = 'mainMember';
-  static const String dependentDtoList = 'dependentDtoList';
-  static const String salesRep = 'salesRep';
-  static const String productDetails = 'productDetails';
-  static const String membershipNumber = 'membershipNumber';
-  static const String invoiceNumber = 'invoiceNumber';
-  static const String membershipPeriod = 'membershipPeriod';
-  static const String creationTime = 'creationTime';
-  static const String passwordStatus = 'passwordStatus';
-  static const String dateJoined = 'dateJoined';
-  static const String partnerId = 'partnerId';
-
-  //CLAIM
-
-  static const String branchCode = 'branchCode';
-  static const String claimantId = 'claimantId';
-  static const String baseUnitOfMeasure = 'baseUnitOfMeasure';
-  static const String sellingPrice = 'sellingPrice';
-  static const String customer = 'customer';
-  static const String salesRepresentative = 'salesRepresentative';
-  static const String createdOn = 'createdOn';
-  static const String price = 'price';
-  static const String surname = 'surname';
-  static const String tombstone = 'tombstone';
-  static const String fileType = 'fileType';
-  static const String no = 'no';
-  static const String fileId = 'fileId';
-  static const String file = 'file';
-  static const String membership = 'membership';
-  static const String paymentMethodDescription = 'paymentMethodDescription';
-  static const String statusDescription = 'statusDescription';
-
-
-
-  static const String employee = 'employee';
-
-  static const String paymentMethod = 'paymentMethod';
-  static const String recipientId = 'recipientId';
-  static const String pricing = 'pricing';
-  static const String gender = 'gender';
-  static const String pricingDescription = 'pricingDescription';
-  static const String documentType = 'documentType';
-  static const String maritalStatus = 'maritalStatus';
-  static const String identity = 'identity';
-  static const String baseUnitOfMeasureDescription = 'baseUnitOfMeasureDescription';
-  static const String uploadDate = 'uploadDate';
-  static const String claims = 'claims';
-  static const String amounts = 'amounts';
-
-  //Notification
-  static const String notificationDtos = 'notificationDtos';
-  static const String SubType = 'SubType';
-
-  //Bank
-  static const String partnerBankAccountDtoList = 'partnerBankAccountDtoList';
-
-  static const String recipient = 'recipient';
-  static const String employeeResponsible = 'employeeResponsible';
-
-
-}
-
-class Resources {
-  static const String address = 'address';
-  static const String bankAccount = 'bank-account';
-  static const String assign = 'assign';
-  static const String terminate = 'terminate';
-  static const String rehire = 'rehire';
-  static const String suspend = 'suspend';
-  static const String validateSAID = 'validateSAID';
-  static const String attachments = 'attachments';
-  static const String otp = 'otp';
-  static const String employee = 'employee';
-  static const String employees = 'employees';
-  static const String authenticate = 'authenticate';
-  static const String tickets = 'tickets';
-  static const String ticketsLog = 'ticketsLog';
-  static const String ticketInProgress = 'inprogress';
-  static const String ticketAwaiting = 'awaitingCustomer';
-  static const String ticketResolve = 'resolved';
-  static const String ticketReject = 'reject';
-  static const String reject = 'reject';
-  static const String dispute = 'dispute';
-  static const String ticketCancel = 'cancel';
-  static const String archive = 'archive';
-  static const String unarchive = 'unarchive';
-  static const String user = 'user';
-  static const String resetPassword = 'passwordreset';
-  static const String reset = 'reset';
-  static const String forgotPassword = 'forgotPassword';
-  static const String versions = 'versions';
-  static const String leaves = 'leaves';
-  static const String fields = 'fields';
-  static const String fieldOptions = 'fieldoptions';
-  static const String leavesApprovers = 'leavesApprovers';
-  static const String approversHistory = 'approversHistory';
-  static const String leavesToApprove = 'leavesToApprove';
-  static const String leaveProfiles = 'leaveProfiles';
-  static const String transactionNotes = 'transactionNotes';
-  static const String persons = 'persons';
-  static const String workcenter = 'workcenter';
-  static const String roleWorkcenters = 'roleWorkcenters';
-  static const String workCenterDesc = 'workcenterDesc';
-  static const String workCenterId = 'id';
-  static const String receipts = 'receipts';
-  static const String receipt = 'receipt';
-  static const String cashup = 'cashup';
-  static const String policies = 'policies';
-
-  static const String voucherValue = 'vouchers';
-  static const String identity = 'identity';
-  static const String contactValue = 'contacts';
-
-  static const String approve = 'approve';
-  static const String approveCancellation = 'ApproveCancellation';
-  static const String approveEditedLeave = 'ApproveEditedLeave';
-  static const String rejectApproval = 'RejectApproval';
-  static const String rejectCancel = 'RejectCancel';
-  static const String rejectEditedLeave = 'rejectEditedLeave';
-  static const String cancel = 'cancel';
-  static const String edit = 'Edit';
-  static const String roles = 'roles';
-  static const String role = 'role';
-  static const String products = 'products';
-  static const String editPosition = 'editPosition';
-  static const String addWorkcenter = 'addWorkcenter';
-  static const String removeWokcenter = 'removeWorkCenter';
-
-  static const String ticketStatusNew = 'new';
-  static const String ticketStatusOpen = 'open';
-  static const String closed = 'closed';
-  static const String inprogress = 'inprogress';
-  static const String awaitingCustomer = 'awaitingCustomer';
-
-  static const String addresses = 'addresses';
-  static const String sendNotifications = 'sendNotifications';
-
-  static const String getTasksAssigned = 'getTasksAssigned';
-  static const String editTask = 'editTask';
-  static const String startTask = 'startTask';
-  static const String cancelTask = 'cancelTask';
-  static const String endTask = 'endTask';
-  static const String completeTask = 'completeTask';
-  static const String getSpecificTask = 'getSpecificTask';
-  static const String getTasks = 'getTasks';
-
-  static const String refreshAuthenticate = 'refreshAuthenticate';
-  static const String comment = 'comment';
-
-  static const String createTask = 'createTask';
-
-  static const String organization = 'Organization';
-  static const String OrganizationUsers = 'OrganizationUsers';
-
-  static const String dependents = 'dependents';
-  static const String claims = 'claims';
-  static const String product = 'product';
-  static const String identities = 'identities';
-  static const String contacts = 'contacts';
-  static const String contact = 'contact';
-  static const String employeeId = 'employeeId';
-  static const String startDate = 'startDate';
-  static const String branch = 'branch';
-  static const String employment = 'employment';
-
-  static const String claimApprove = 'approve';
-  static const String claimCancel = 'cancel';
-  static const String claimDecline = 'decline';
-
-  static const String deposits = 'deposits';
-  static const String deposit = 'deposit';
-
-  static const String attributes = 'attributes';
-  static const String pricing = 'pricing';
-  static const String refer = 'refer';
-  static const String vouchers = 'vouchers';
-
-  static const String submit = 'submit';
-  static const String submitted = 'submitted';
-
-  //KEYS FOR THE NEW BACKEND
-
-  static const String purchaseOrder = 'purchase-order';
-  static const String salesOrder = 'sales-order';
-  static const String invoice = 'invoice';
-  static const String field = 'field';
-  static const String option = 'option';
-  static const String claim = 'claim';
-  static const String membership = 'membership';
-  static const String dependent = 'dependent';
-  static const String partner = 'partner';
-
-  static const String quotation = 'quotation';
-  static const String items = 'items';
-
-  static const String calculate_pricing = 'calculate-pricing';
-  static const String application = 'application';
-  static const String customer = 'customer';
-  static const String prospects = 'prospects';
-  static const String prospect = 'prospect';
-  static const String supplier = 'supplier';
-  static const String paymentRequest = 'payment-request';
-
-  static const String attachmentfile = 'attachmentfile';
-  static const String attachment = 'attachment';
-
-  static const String lock = 'lock';
-  static const String unlock = 'unlock';
-  static const String byId = 'byId';
-  static const String addressId = 'addressId';
-  static const String validate = 'validate';
-
-  static const String layBy = 'layBy';
-  static const String voucher = 'voucher';
-  static const String transactionLink = 'transactionLink';
-  static const String transaction = 'transaction';
-  static const String serviceRequest = 'service-request';
-  static const String tombstoneRecipient = 'tombstone-recipient';
-  static const String attribute = 'attribute';
-  static const String premium = 'premium';
-  static const String booking = 'booking';
-  static const String groupSociety = 'group-society';
-  static const String premiums = 'premiums';
-
-  static const String notifications = 'notifications';
-  static const String read = 'read';
-  static const String byTrans = 'byTrans';
-
-  static const String complete = 'complete';
-
-
-
-}
-
-class JsonPayloads {
-  static const String suburb = 'suburb';
-  static const String city = 'city';
-  static const String town = 'town';
-  static const String province = 'province';
-
-  static const String claimId = 'claimId';
-  static const String comments = 'comments';
-  static const String id = 'id';
-  static const String filter = 'filter';
-  static const String filterValue = 'x';
-  static const String assignedTo = 'assignedTo';
-  static const String assignedToID = 'assignedToID';
-  static const String clintID = 'clintID';
-  static const String status = 'status';
-
-  static const String partnerID = 'partnerID';
-  static const String ticketID = 'ticketID';
-
-  static const String value = 'value';
-  static const String type = 'type';
-  static const String transaction = 'transaction';
-
-  static const String otp = 'otp';
-  static const String partnerEmail = 'partnerEmail';
-
-
-  static const String number = 'number';
-  static const String New = 'New';
-  static const String Open = 'Open';
-  static const String Closed = 'Closed';
-  static const String InProgress = 'Inprogress';
-  static const String Completed = 'Completed';
-  static const String Resolved = 'Resolved';
-
-  static const String loggedByID = 'loggedByID';
-  static const String approverID = 'approverID';
-  static const String startDate = 'startDate';
-  static const String endDate = 'endDate';
-  static const String description = 'description';
-  static const String subDescription = 'subDescription';
-  static const String leaveType = 'leaveType';
-
-  static const String documentType = 'documentType';
-  static const String parentType = 'parentType';
-  static const String parentReference = 'parentReference';
-
-  static const String fieldCode = 'code';
-  static const String fieldDescription = 'description';
-
-  static const String fieldOptionCode = 'code';
-  static const String fieldOptionDescription = 'description';
-
-  static const String field = 'field';
-
-  static const String workCenter = 'workcenter';
-  static const String position = 'position';
-
-  static const String workCenterDesc = 'description';
-  static const String workCenterPath = 'path';
-
-  static const String appName = 'appName';
-  static const String versionNumber = 'versionNumber';
-  static const String appUsable = 'appUsable';
-
-  static const String email = 'email';
-  static const String detail = 'detail';
-  static const String cellphone = 'cellphone';
-  static const String role = 'role';
-  static const String password = 'password';
-  static const String userType = 'userType';
-  static const String partner = 'partner';
-
-  static const String ticketDescription = 'description';
-  static const String ticketCategory = 'category';
-  static const String ticketSummary = 'subDescription';
-  static const String ticketPriority = 'priority';
-
-  static const String recievedBy = 'recievedBy';
-
-  static const String taskDescription = 'taskDescription';
-  static const String dueDate = 'dueDate';
-  static const String assignedBy = 'assignedBy';
-  static const String summary = 'summary';
-
-  static const String sms = 'sms';
-
-  static const String ticketTaskID = 'ticketTaskID';
-
-  static const String firstName = 'firstName';
-  static const String middleName = 'middleName';
-  static const String lastName = 'lastName';
-  static const String idNumber = 'idNumber';
-  static const String personIdNumber = 'idNumber';
-  static const String idType = 'idType';
-  static const String addressline1 = 'line1';
-  static const String addressline2 = 'line2';
-  static const String addressline3 = 'line3';
-  static const String addressline4 = 'line4';
-  static const String birthDate = 'birthDate';
-  static const String postalCode = 'postalCode';
-  static const String cellNumber = 'detail';
-  static const String gender = 'gender';
-  static const String title = 'title';
-  static const String maritalStatus = 'maritalStatus';
-
-  static const String validTo = 'validTo';
-  static const String attribute = 'attribute';
-  static const String validFrom = 'validFrom';
-  static const String priceType = 'priceType';
-  static const String priceTypeDescription = 'priceTypeDescription';
-  static const String notes = 'notes';
-  static const String declineReason = 'declineReason';
-
-  static const String productCategory = 'category';
-  static const String productDescription = 'description';
-  static const String quantity = 'quantity';
-
-  static const String deceasedId = 'deceasedId';
-  static const String claimantId = 'claimantId';
-  static const String bankAccount = 'bankAccount';
-  static const String bankName = 'bankName';
-  static const String branchName = 'branchName';
-  static const String branchCode = 'branchCode';
-  static const String accountType = 'accountType';
-  static const String accountNumber = 'accountNumber';
-
-  static const String categoryDescription = 'categoryDescription';
-
-  //KEYS FOR THE NEW BACKEND
-  static const String customer = 'customer';
-  static const String items = 'items';
-  static const String product = 'product';
-  static const String unit = 'items';
-  static const String unitPrice = 'unitPrice';
-  static const String deliveryDate = 'deliveryDate';
-  static const String supplier = 'supplier';
-  static const String claimant = 'claimant';
-  static const String deceased = 'deceased';
-  static const String member = 'member';
-  static const String membership = 'membership';
-  static const String deathDate = 'deathDate';
-  static const String burialDate = 'burialDate';
-  static const String code = 'code';
-  static const String category = 'category';
-  static const String baseUnitOfMeasure = 'baseUnitOfMeasure';
-  static const String sellingPrice = 'sellingPrice';
-
-  static const String customerId = 'customerId';
-  static const String expiryDate = 'expiryDate';
-  static const String productId = 'productId';
-  static const String ean = 'ean';
-  static const String uom = 'uom';
-  static const String lineTotal = 'lineTotal';
-
-  static const String username = 'username';
-  static const String amountDeposited = 'amountDeposited';
-
-  static const String name1 = 'name1';
-  static const String name2 = 'name2';
-  static const String name3 = 'name3';
-  static const String name4 = 'name4';
-
-  static const String employeeResponsibleId = 'employeeResponsibleId';
-  static const String reference = 'reference';
-  static const String amount = 'amount';
-  static const String bankDetails = 'bankDetails';
-
-  static const String file = 'file';
-  static const String uploadDate = 'uploadDate';
-  static const String uploadTime = 'uploadTime';
-  static const String uploadedBy = 'uploadedBy';
-  static const String downloadDate = 'downloadDate';
-  static const String downloadedBy = 'downloadedBy';
-  static const String reason = 'reason';
-
-  static const String previousProduct = 'previousProduct';
-  static const String statusReason = 'statusReason';
-  static const String salesRepresentativeId = 'salesRepresentativeId';
-  static const String premium = 'premium';
-
-  static const String line1 = 'line1';
-  static const String line2 = 'line2';
-  static const String line3 = 'line3';
-  static const String line4 = 'line4';
-
-  static const String employeeId = 'employeeId';
-  static const String branch = 'branch';
-  static const String groupID = 'groupID';
-  static const String department = 'department';
-  static const String identityDtos = 'identityDtos';
-  static const String addressDtos = 'addressDtos';
-  static const String contactDtos = 'contactDtos';
-
-  static const String partnerType = 'partnerType';
-  static const String surname = 'surname';
-  static const String organizationName = 'organizationName';
-
-  static const String memberId = 'memberId';
-  static const String membershipId = 'membershipId';
-  static const String dateOfBirth = 'dateOfBirth';
-  static const String membershipNumber = 'membershipNumber';
-  static const String receiptType = 'receiptType';
-  static const String invoiceNumber = 'invoiceNumber';
-  static const String membershipPeriod = 'membershipPeriod';
-  static const String tenderType = 'tenderType';
-  static const String salesRepresentative = 'salesRepresentative';
-  static const String salesArea = 'salesArea';
-  static const String dateJoined = 'dateJoined';
-  static const String price = 'price';
-  static const String pricingType = 'pricingType';
-  static const String transactionIdLink = 'transactionIdLink';
-  static const String partnerId = 'partnerId';
-  static const String accountHolder = 'accountHolder';
-  static const String layById = 'layById';
-  static const String fileType = 'fileType';
-  static const String autoGenerateCode = 'autoGenerateCode';
-  static const String creationType = 'creationType';
-  static const String previousInsurerId = 'previousInsurerId';
-  static const String pricing = 'pricing';
-  static const String lastReceiptDate = 'lastReceiptDate';
-  static const String membershipType = 'membershipType';
-  static const String recipientId = 'recipientId';
-  static const String paymentReason = 'paymentReason';
-  static const String paymentMethod = 'paymentMethod';
-  static const String location = 'location';
-  static const String terminalId = 'terminalId';
-  static const String objectType = 'objectType';
-  static const String objectId = 'objectId';
-  static const String attributeValue = 'value';
-  static const String bookDate = "bookDate";
-  static const String bookTime = "bookTime";
-  static const String receipts = 'receipts';
-  static const String dateAdded = 'dateAdded';
-  static const String dateEffective = 'dateEffective';
-  static const String language = 'language';
-
-  static const String processor = 'processor';
-  static const String recipient = 'recipient';
-  static const String subType = 'subType';
-  static const String extension = 'extension';
-
-
-// static const String  = '';
-}
-
-class QueryParameters {
-  static const String transactionId = 'transactionId';
-
-  static const String versionAppName = 'appName';
-  static const String versionAppUsable = 'appUsable';
-  static const String approverId = 'approverId';
-  static const String assignedTo = 'assignedTo';
-  static const String assignedToID = 'assignedToID';
-
-  static const String category = 'category';
-  static const String clintID = 'clintID';
-
-  static const String documentType = 'documentType';
-
-  static const String endDAte = 'endDate';
-  static const String extension = 'extension';
-
-  static const String field = 'field';
-  static const String filter = 'filter';
-  static const String filterValue = 'x';
-  static const String userType = 'userType';
-
-  static const String id = 'id';
-
-  static const String loggedByID = 'loggedByID';
-  static const String loggedById = 'loggedById';
-
-  static const String personIdNumber = 'idumber';
-  static const String partnerRole = 'partnerRole';
-  static const String idNumber = 'idNumber';
-  static const String idType = 'idType';
-
-  static const String organisationId = 'organisationId';
-  static const String otp = 'otp';
-
-  static const String parentType = 'parentType';
-  static const String parentReference = 'parentReference';
-  static const String partnerFunctionEmployee = 'EMPLOYEE';
-  static const String partnerFunctionOrganization = 'ORGANIZATION';
-  static const String partnerFunction = 'partnerFunction';
-  static const String partnerID = 'partnerID';
-  static const String partnerId = 'partnerId';
-  static const String partnerNo = 'partnerNo';
-  static const String processedBy = 'processedBy';
-  static const String processor = 'processor';
-
-  static const String reference = 'reference';
-
-  static const String status = 'status';
-
-  static const String ticketID = 'ticketID';
-  static const String transaction = 'transaction';
-
-  static const String versionApkVersionCode = 'versionNumber';
-  static const String value = 'value';
-
-  static const String meesageType = 'meesageType';
-  static const String adminEmail = 'adminEmail';
-  static const String user = 'user';
-  static const String password = 'password';
-
-  static const String assignTo = 'assignTo';
-  static const String ticketTaskID = 'ticketTaskID';
-  static const String summary = 'summary';
-  static const String taskDescription = 'taskDescription';
-  static const String dueDate = 'dueDate';
-  static const String taskLogID = 'taskLogID';
-
-  static const String userID = 'userID';
-  static const String note = 'note';
-  static const String type = 'type';
-
-  static const String workcenter = 'workcenter';
-  static const String position = 'position';
-
-  static const String New = 'new';
-  static const String inprogress = 'inprogress';
-  static const String open = 'open';
-  static const String cancelled = 'cancelled';
-  static const String resolved = 'resolved';
-  static const String completed = 'completed';
-
-  static const String organizationId = 'organizationId';
-
-  static const String policy = 'policy';
-  static const String approver = 'approver';
-
-  static const String salesArea = 'salesArea';
-  static const String product = 'product';
-  static const String customer = 'customer';
-  static const String salesRepresentative = 'salesRepresentative';
-
-  static const String amount = 'amount';
-  static const String location = 'location';
-  static const String terminalType = 'terminalType';
-  static const String terminalId = 'terminalId';
-  static const String tenderType = 'tenderType';
-  static const String transactionLink = 'transactionLink';
-  static const String role = 'role';
-
-  static const String cashupId = 'cashupId';
-  static const String checkoutId = 'checkoutId';
-  static const String statusReason = 'statusReason';
-  static const String description = 'description';
-
-  static const String email = 'email';
-  static const String cellphone = 'cellphone';
-  static const String passwordStatus = 'passwordStatus';
-  static const String userRole = 'userRole';
-
-  static const String fieldOption = 'fieldOption';
-
-  static const String addressId = 'addressId';
-  static const String startDate = 'startDate';
-  static const String createdBy = 'createdBy';
-  static const String createdOn = 'createdOn';
-  static const String customerId = 'customerId';
-  static const String endDate = 'endDate';
-  static const String number = 'number';
-
-  static const String membershipPeriod = 'membershipPeriod';
-  static const String membershipNumber = 'membershipNumber';
-  static const String invoiceNumber = 'invoiceNumber';
-  static const String receiptType = 'receiptType';
-  static const String transactionLinkId = 'transactionLinkId';
-  static const String fileType = 'fileType';
-  static const String claimant = 'claimant';
-  static const String no = 'no';
-  static const String deceased = 'deceased';
-  static const String member = 'member';
-  static const String membership = 'membership';
-  static const String deathDate = 'deathDate';
-  static const String burialDate = 'burialDate';
-
-  static const String contactType = 'contactType';
-  static const String attribute = 'attribute';
-  static const String notCashed = 'notCashed';
-  static const String pricing = 'pricing';
-  static const String employeeResponsibleId = 'employeeResponsibleId';
-  static const String receipts = 'receipts';
-  static const String employeeId = "employeeId";
-  static const String bookDate = "bookDate";
-  static const String attributeName = "attributeName";
-  static const String attributeValue = "attributeValue";
-  static const String objectId = "objectId";
-
-// static const String  = '';
-}
-
-class Statuses {
-  static const String pending = 'Pending';
-  static const String approved = 'Approved';
-  static const String cancelled = 'Cancelled';
-  static const String rejected = 'Rejected';
-}
-
-class StatusReasons {
-  static const String awaitingApproval = 'Awaiting Approval';
-  static const String awaitingCancelApproval = 'Awaiting Cancel Approval';
-  static const String awaitingEditApproval = 'Awaiting Edit Approval';
-  static const String Rejected = 'Rejected';
-}
-
-class SharedPrefs {
-  static const server = 'server';
-  static const token = 'token';
-  static const username = 'username';
-  static const lastPage = 'lastPage';
-  static const startTime = 'startTime';
-  static const endTime = 'endTime';
-  static const isTracking = 'isTracking';
-  static const trackingTicketID = 'trackingTicketID';
-  static const refreshToken = 'refreshToken';
-  static const initialRoute = 'initialRoute';
-  static const loginScreenID = 'loginScreenID';
-  static const loginRole = 'loginRole';
-  static const isLoggedIn = 'isLoggedIn';
-  static const tenantID = 'tenantID';
-  static const userID = 'userID';
-  static const role = 'role';
-  static const workcenters = 'workcenters';
-  static const loggedInUser = 'loggedInUser';
-  static const roleRoot = 'roleRoot';
-}
-
-class CommentTypes {
-  static const String ticketComment = 'TICKETCOMMENT';
-  static const String leaveComment = 'LEAVEMANAGEMENT';
-}
-
-class FieldOptionTypes {
-  static const String fieldOptionDisputeReason = 'CLAIM-DISPUTE-REASON';
-  static const String fieldOptionCategory = 'CATEGORY';
-  static const String fieldOptionPriority = 'PRIORITY';
-  static const String fieldOptionIdType = 'ID-TYPE';
-  static const String fieldOptionGender = 'GENDER';
-  static const String fieldOptionTitle = 'TITLE';
-  static const String fieldOptionMaritalStatus = 'MARITAL-STATUS';
-  static const String fieldOptionSuburb = 'SUBURB';
-  static const String fieldOptionTown = 'TOWN';
-  static const String fieldOptionCity = 'CITY';
-  static const String fieldOptionContactType = 'CONTACT-TYPE';
-  static const String fieldAddress = 'ADDRESSES';
-  static const String fieldOptionAddressType = 'ADDRESS-TYPE';
-  static const String fieldPartnerRole = 'PARTNER-ROLE';
-  static const String fieldClaimDeclineReason = 'CLAIM-DECLINE-REASON';
-  static const String fieldPaymentMethod = 'PAYMENT-METHOD';
-  static const String fieldOptionPaymentRequestStatusReason = 'PAYMENT-REQUEST-STATUS-REASON';
-  static const String fieldPaymentRequestType = 'PAYMENT-REQUEST-TYPE';
-  static const String fieldBankAccountType = 'BANK-ACCOUNT-TYPE';
-  static const String fieldPaymentReason = 'PAYMENT-REASON';
-  static const String fieldBankName = 'BANK-NAME';
-  static const String fieldBranch = 'BRANCH';
-  static const String fieldDepartment = 'DEPARTMENT';
-  static const String fieldEmploymentType = 'EMPLOYMENT-TYPE';
-  static const String fieldJobTitle = 'JOB-TITLE';
-
-  static const String fieldClaimTypes = 'CLAIM-TYPE';
-  static const String fieldOptionBusinessIdentity = 'BUSINESS-IDENTITIES';
-
-  static const String productCategory = 'PRODUCT-CATEGORY';
-  static const String fieldOptionLayByStatus = 'LAYBY-STATUS';
-  static const String fieldOptionLayByStatusReason = 'LAYBY-STATUS-REASON';
-  static const String fieldOptionPartnerType = 'PARTNER-TYPE';
-  static const String uom = 'UOM';
-  static const String pricingType = 'PRICING-TYPE';
-  static const String membershipCreationType = 'MEMBERSHIP-CREATION-TYPE';
-  static const String suburb = 'SUBURB';
-  static const String userType = 'USER-TYPE';
-  static const String productAttribute = 'PRODUCT-ATTRIBUTE';
-  static const String membershipType = 'MEMBERSHIP-TYPE';
-  static const String salesArea = 'SALES-AREA';
-  static const String tenderType = 'TENDER-TYPE';
-  static const String documentType = 'DOCUMENT-TYPE';
-  static const String membershipStatuses = 'MEMBERSHIP-STATUSES';
-  static const String membershipStatusReasons = 'MEMBERSHIP-STATUS-REASONS';
-}
-
-class PathNames {
-  static const String identityPath = 'Identity';
-  static const String rolePath = 'Role';
-  static const String contactPath = 'Contact';
-  static const String addressPath = 'Address';
-  static const String employmentPath = 'Employment Details';
-  static const String employeePath = 'Employee Details';
-  static const String customerPath = 'Customer Details';
-}
-
-class PartnerTypes {
-  static const String dependent = 'DEPENDENT';
-  static const String customer = 'CUSTOMER';
-  static const String employee = 'EMPLOYEE';
-  static const String supplier = 'SUPPLIER';
-  static const String organization = 'ORGANIZATION';
-  static const String organisation = 'ORGANISATION';
-  static const String person = 'PERSON';
-  static const String group = 'GROUP';
-  static const String individual = 'individual';
-}
-
-class PartnerRoles {
-  static const String competitor = 'COMPETITOR';
-  static const String dependent = 'DEPENDENT';
-  static const String customer = 'CUSTOMER';
-  static const String employee = 'EMPLOYEE';
-}
-
-
-class PartnerAttributes {
-  static const String identity = 'identity';
-  static const String contact = 'contact';
-  static const String address = 'address';
-  static const String bank = 'bank';
-}
-
-class MembershipTypes {
-  static const String group = 'GROUP';
-  static const String individual = 'INDIVIDUAL';
-}
-
-class ProductTypes {
-  static const String groupSociety = 'GROUP-SOCIETY';
-  static const String membership = 'MEMBERSHIP';
-}
-
-class ObjectTypes {
-  static const String transaction = 'TRANSACTION';
-  static const String partner = 'PARTNER';
-}
